requirements.txt
json5
pathspec
pytest
pyyaml
